+++ apps/frontend/src/libs/enums/enums.ts
@@ -10,12 +10,14 @@ export { LoaderIconConfig } from "./loader-icon-config.enum.js";
 export { LoaderProgressConfig } from "./loader-progress-config.enum.js";
 export { LogoSize } from "./logo-size.enum.js";
 export { LogoTheme } from "./logo-theme.enum.js";
+export { LogoType } from "./logo-type.enum.js";
 export { NotificationMessage } from "./notification-massage.enum.js";
 export { SearchDebounceDelay } from "./search-debounce-delay.enum.js";
 export {
 	APIPath,
 	AppEnvironment,
 	ContentType,
+	HTTPMethod,
 	KeyboardKey,
 	MeetingHost,
 	ServerErrorType,