+++ packages/shared/src/modules/meetings/libs/enums/meetings-api-path.enum.ts
@@ -1,5 +1,8 @@
 const MeetingsApiPath = {
 	$ID: "/:id",
+	$ID_PUBLIC: "/:id/public",
+	$ID_STOP_RECORDING: "/:id/stop-recording",
+	$ID_URL: "/:id/url",
 	ROOT: "/",
 } as const;