+++ packages/shared/src/modules/meetings/meetings.ts
@@ -6,7 +6,9 @@ export {
 } from "./libs/enums/enums.js";
 export {
 	type MeetingCreateRequestDto,
+	type MeetingDetailedResponseDto,
 	type MeetingGetAllResponseDto,
+	type MeetingGetPublicUrlResponseDto,
 	type MeetingResponseDto,
 	type MeetingUpdateRequestDto,
 } from "./libs/types/types.js";